refactor interface in 2.json_rpc
