+++ course4/1.RPC/2.json_rpc/task4.1.2.1/internal/controller/router.go
@@ -10,6 +10,15 @@ import (
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
 
+type Controller interface {
+	Register(w http.ResponseWriter, r *http.Request)
+	Auth(w http.ResponseWriter, r *http.Request)
+	Search(w http.ResponseWriter, r *http.Request)
+	Geocode(w http.ResponseWriter, r *http.Request)
+	GetByEmail(w http.ResponseWriter, r *http.Request)
+	GetAllUsers(w http.ResponseWriter, r *http.Request)
+}
+
 func NewRouter(controller Controller, auth *jwtauth.JWTAuth) *chi.Mux {
 	router := chi.NewRouter()
 	router.Use(middleware.Logger)
