+++ course4/1.RPC/2.json_rpc/task4.1.2.1/internal/repository/repository.go
@@ -19,13 +19,6 @@ type PostgressDataBase struct {
 	metrics *metrics.ProxyMetrics
 }
 
-type UserRepository interface {
-	Create(ctx context.Context, user models.User) error
-	GetByID(ctx context.Context, id string) (models.User, error)
-	GetByEmail(ctx context.Context, email string) (models.User, error)
-	List(ctx context.Context) ([]models.User, error)
-}
-
 func StartPostgressDataBase(ctx context.Context, connStr string) (*PostgressDataBase, error) {
 	dataBase := &PostgressDataBase{}
 	db, err := sql.Open("postgres", connStr)
