+++ course4/1.RPC/2.json_rpc/task4.1.2.1/internal/service/geoRPC.go
@@ -6,11 +6,6 @@ import (
 	"metrics/internal/models"
 )
 
-type GeodataServiceRPC interface {
-	Search(geocode models.RequestAddressGeocode) (models.ResponseAddress, error)
-	Geocode(address models.ResponseAddress) (models.ResponseAddressGeocode, error)
-}
-
 type GeoRPC struct {
 	client *rpc.Client
 }
