+++ course4/1.RPC/2.json_rpc/task4.1.2.1/internal/service/user_service.go
@@ -17,19 +17,25 @@ import (
 
 	"metrics/internal/metrics"
 	"metrics/internal/models"
-	"metrics/internal/repository"
 )
 
-type UserService interface {
-	CreateUser(user models.User) error
-	AuthUser(user models.User) (string, error)
-	GetUserByEmail(email string) (models.User, error)
-	GetAllUsers() ([]models.User, error)
+type UserRepository interface {
+	Create(ctx context.Context, user models.User) error
+	GetByID(ctx context.Context, id string) (models.User, error)
+	GetByEmail(ctx context.Context, email string) (models.User, error)
+	List(ctx context.Context) ([]models.User, error)
 }
 
 type UserServiceImpl struct {
-	repo  repository.UserRepository
+	repo  UserRepository
 	token *jwtauth.JWTAuth
+}
+
+func NewUserServiceImpl(repo UserRepository, token *jwtauth.JWTAuth) *UserServiceImpl {
+	return &UserServiceImpl{
+		repo:  repo,
+		token: token,
+	}
 }
 
 func (s *UserServiceImpl) CreateUser(user models.User) error {
@@ -67,17 +73,25 @@ func (s *UserServiceImpl) GetUserByEmail(email string) (models.User, error) {
 	return s.repo.GetByEmail(context.Background(), email)
 }
 
-func NewUserServiceImpl(repo repository.UserRepository, token *jwtauth.JWTAuth) UserService {
-	return &UserServiceImpl{
-		repo:  repo,
-		token: token,
-	}
+type UserService interface {
+	CreateUser(user models.User) error
+	AuthUser(user models.User) (string, error)
+	GetUserByEmail(email string) (models.User, error)
+	GetAllUsers() ([]models.User, error)
 }
 
 type UserServiceProxy struct {
 	userService UserService
 	client      *redis.Client
 	metrics     *metrics.ProxyMetrics
+}
+
+func NewUserServiceProxy(userService UserService, client *redis.Client) *UserServiceProxy {
+	return &UserServiceProxy{
+		userService: userService,
+		client:      client,
+		metrics:     metrics.NewProxyMetrics(),
+	}
 }
 
 func (s *UserServiceProxy) CreateUser(user models.User) error {
@@ -134,12 +148,4 @@ func (s *UserServiceProxy) GetUserByEmail(email string) (models.User, error) {
 	user := models.User{}
 	err = json.Unmarshal([]byte(data), &user)
 	return user, err
-}
-
-func NewUserServiceProxy(userService UserService, client *redis.Client) UserService {
-	return &UserServiceProxy{
-		userService: userService,
-		client:      client,
-		metrics:     metrics.NewProxyMetrics(),
-	}
 }
